Fix "remember tfa" cookie hijacking
The `remember_tfa` cookie can be re-used even after its expirations
date by simply copying the cookie. This commit improves the security
of the cookie by adding the expiration date to the cookie itself and
validating the expiration date before accepting it.
It also allows defining a `remember_tfa_token` method in the resource
that may be used for invalidating all cookies.